# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DocumentContent-Input": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "items": {
              "maxLength": 20000,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10000,
            "minItems": 1,
            "title": "Body",
            "type": "array"
          },
          "doc_type": {
            "pattern": "^[a-z][a-z0-9_]{0,49}$",
            "title": "Doc Type",
            "type": "string"
          },
          "requisites": {
            "maxProperties": 30,
            "patternProperties": {
              "^[a-z][a-z0-9_]{0,49}$": {
                "maxLength": 500,
                "type": "string"
              }
            },
            "title": "Requisites",
            "type": "object"
          }
        },
        "required": [
          "doc_type",
          "body"
        ],
        "title": "DocumentContent",
        "type": "object"
      },
      "DocumentContent-Output": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "items": {
              "maxLength": 20000,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10000,
            "minItems": 1,
            "title": "Body",
            "type": "array"
          },
          "doc_type": {
            "pattern": "^[a-z][a-z0-9_]{0,49}$",
            "title": "Doc Type",
            "type": "string"
          },
          "requisites": {
            "maxProperties": 30,
            "patternProperties": {
              "^[a-z][a-z0-9_]{0,49}$": {
                "maxLength": 500,
                "type": "string"
              }
            },
            "title": "Requisites",
            "type": "object"
          }
        },
        "required": [
          "doc_type",
          "requisites",
          "body"
        ],
        "title": "DocumentContent",
        "type": "object"
      },
      "DocumentFormatting": {
        "additionalProperties": false,
        "description": "Validated Word-like overrides applied to the prepared document on download.",
        "properties": {
          "body_alignment": {
            "enum": [
              "left",
              "center",
              "right",
              "justify"
            ],
            "title": "Body Alignment",
            "type": "string"
          },
          "body_bold": {
            "title": "Body Bold",
            "type": "boolean"
          },
          "body_italic": {
            "title": "Body Italic",
            "type": "boolean"
          },
          "body_underline": {
            "title": "Body Underline",
            "type": "boolean"
          },
          "first_line_indent_mm": {
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "First Line Indent Mm",
            "type": "number"
          },
          "font": {
            "enum": [
              "Times New Roman",
              "Arial",
              "Calibri",
              "Georgia",
              "Courier New"
            ],
            "title": "Font",
            "type": "string"
          },
          "font_size": {
            "maximum": 20.0,
            "minimum": 10.0,
            "title": "Font Size",
            "type": "number"
          },
          "line_spacing": {
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Line Spacing",
            "type": "number"
          },
          "paragraph_space_after_pt": {
            "maximum": 24.0,
            "minimum": 0.0,
            "title": "Paragraph Space After Pt",
            "type": "number"
          }
        },
        "required": [
          "font",
          "font_size",
          "line_spacing",
          "paragraph_space_after_pt",
          "first_line_indent_mm",
          "body_alignment",
          "body_bold",
          "body_italic",
          "body_underline"
        ],
        "title": "DocumentFormatting",
        "type": "object"
      },
...
      "DownloadRequest": {
        "additionalProperties": false,
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentContent-Input"
          },
          "formatting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentFormatting"
              },
              {
                "type": "null"
              }
            ]
          },
          "template_id": {
            "pattern": "^[a-z][a-z0-9_]{0,49}$",
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "document",
          "template_id"
        ],
        "title": "DownloadRequest",
        "type": "object"
      },
      "ErrorInfo": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorIssue"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "request_id",
          "details"
        ],
        "title": "ErrorInfo",
        "type": "object"
      },
      "ErrorIssue": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ErrorIssue",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "required": [
          "detail",
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ProcessRequest": {
        "additionalProperties": false,
        "properties": {
          "doc_type": {
            "pattern": "^[a-z][a-z0-9_]{0,49}$",
            "title": "Doc Type",
            "type": "string"
          },
          "draft": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Draft",
            "type": "string"
          },
          "requisites": {
            "maxProperties": 30,
            "patternProperties": {
              "^[a-z][a-z0-9_]{0,49}$": {
                "maxLength": 500,
                "type": "string"
              }
            },
            "title": "Requisites",
            "type": "object"
          }
        },
        "required": [
          "doc_type",
          "draft"
        ],
        "title": "ProcessRequest",
        "type": "object"
      },
      "ProcessResponse": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "items": {
              "type": "string"
            },
            "title": "Changes",
            "type": "array"
          },
          "document": {
            "$ref": "#/components/schemas/DocumentContent-Output"
          },
          "missing_fields": {
            "items": {
              "$ref": "#/components/schemas/RequisiteField"
            },
            "title": "Missing Fields",
            "type": "array"
          },
          "processor_mode": {
            "title": "Processor Mode",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "document",
          "missing_fields",
          "changes",
          "warnings",
          "processor_mode"
        ],
        "title": "ProcessResponse",
        "type": "object"
      },
...
      "RequisiteField": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "pattern": "^[a-z][a-z0-9_]{0,49}$",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "placement": {
            "default": "labeled",
            "enum": [
              "labeled",
              "letterhead",
              "addressee",
              "registration",
              "headline",
              "salutation",
              "paragraph",
              "signature",
              "executor"
            ],
            "title": "Placement",
            "type": "string"
          },
          "prefix": {
            "default": "",
            "title": "Prefix",
            "type": "string"
          },
          "required": {
            "default": false,
            "title": "Required",
            "type": "boolean"
          },
          "summary": {
            "default": false,
            "title": "Summary",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "label",
          "required",
          "placement",
          "prefix",
          "summary"
        ],
        "title": "RequisiteField",
        "type": "object"
      },
      "RequisiteSuggestions": {
        "additionalProperties": false,
        "properties": {
          "requisites": {
            "maxProperties": 30,
            "patternProperties": {
              "^[a-z][a-z0-9_]{0,49}$": {
                "maxLength": 500,
                "type": "string"
              }
            },
            "title": "Requisites",
            "type": "object"
          }
        },
        "required": [
          "requisites"
        ],
        "title": "RequisiteSuggestions",
        "type": "object"
      },
      "SuggestRequest": {
        "additionalProperties": false,
        "description": "Черновик и тип: что из него можно подставить в форму реквизитов.",
        "properties": {
          "doc_type": {
            "pattern": "^[a-z][a-z0-9_]{0,49}$",
            "title": "Doc Type",
            "type": "string"
          },
          "draft": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Draft",
            "type": "string"
          }
        },
        "required": [
          "doc_type",
          "draft"
        ],
        "title": "SuggestRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend команды: подготовка содержания и независимый экспорт DOCX.",
    "title": "Документ за 3 шага",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/documents/download": {
      "post": {
        "description": "Render user-supplied content without calling the model or claiming fact verification.",
        "operationId": "downloadDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Editable DOCX"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input or catalog selection"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal processing error"
          }
        },
        "summary": "Download",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/process": {
      "post": {
        "description": "Prepare content once; template selection is intentionally absent from this request.",
        "operationId": "processDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input or catalog selection"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal processing error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "LLM returned unusable content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Processing unavailable or busy"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "LLM timed out"
          }
        },
        "summary": "Process",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/requisites/suggest": {
      "post": {
        "description": "A form hint, not processing: no model is involved and its failures cannot reach it.",
        "operationId": "suggestRequisites",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequisiteSuggestions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input or catalog selection"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal processing error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "LLM returned unusable content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Processing unavailable or busy"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "LLM timed out"
          }
        },
        "summary": "Suggest",
        "tags": [
          "documents"
        ]
      }
    }
  }
}